// src/types/snake.ts
export type Position = {
  x: number;
  y: number;
};

export type Direction = 'UP' | 'DOWN' | 'LEFT' | 'RIGHT';

export type GameState = {
  snake: Position[];
  food: Position;
  direction: Direction;
  isGameOver: boolean;
  score: number;
};

export type GameConfig = {
  gridSize: number;
  cellSize: number;
  initialSpeed: number;
}; 